windows: Add some casts to avoid compiler warnings

+++ src/brickd/bricklet_stack.c
@@ -180,7 +180,7 @@ static uint16_t bricklet_stack_check_missing_length(BrickletStack *bricklet_stac
 			return 0;
 		}
 
-		return ret;
+		return (uint16_t)ret;
 	}
 
 	return 0;
